🧪 test(web): rocketAnimationView
